+++ test/modules/Auction/MockAtomicAuctionModule.sol
@@ -86,7 +86,7 @@ contract MockAtomicAuctionModule is AuctionModule {
         revert Auction_NotImplemented();
     }
 
-    function _refundBid(uint96, uint64, address) internal virtual override returns (uint256) {
+    function _refundBid(uint96, uint64, address) internal virtual override returns (uint96) {
         revert Auction_NotImplemented();
     }
